requirements.txt
# MuseAI - Neural Style Transfer
# Core ML Framework
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
Pillow>=9.0.0
opencv-python>=4.7.0

# Face Detection & Recognition
facenet-pytorch>=2.5.3
mtcnn>=0.1.1

# Perceptual Metrics
lpips>=0.1.4
pytorch-msssim>=1.0.0

# Web Interface
gradio>=4.0.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
matplotlib>=3.7.0
scikit-image>=0.21.0

# Configuration
pyyaml>=6.0
